css-align-3_dev/xhtml1print/place-content-shorthand-005.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Box Alignment: place-content shorthand - inherit value</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez" />
<link href="http://www.w3.org/TR/css3-align/#propdef-place-content" rel="help" />
<meta content="Check that place-content's 'inherit' value expands to 'align-content' and 'justify-content'." name="assert" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        place-content: start end;
    }
</style>
</head><body><div id="log"></div>
<div id="test"></div>
<script>
    var child = document.createElement("div");
    document.getElementById("test").appendChild(child);
    child.setAttribute("style", "place-content: inherit");
    var style = getComputedStyle(child);

    test(function() {
        assert_equals(style.getPropertyValue("align-content"),
            "start", "place-content resolved value for align-content");
    }, "Check place-content: inherit - align-content resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-content"),
           "end", "place-content resolved value for justify-content");
    }, "Check place-content: inherit - justify-content resolved value");
</script>
</body></html>
